BugFix
git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@1842 88ce85bb-8564-4ae6-884f-ab7503230054

+++ simulation/grvascular.cpp
@@ -24,7 +24,7 @@ Vascular::~Vascular()
 
 void Vascular::solveVascularSources(GrGrid& grid, double dt, const int time, int DiffStep)
 {
-    int realTime = time*600 + DiffStep*_PARAM(_timestepDiffusion);
+    int realTime = static_cast<int>(time*600 + DiffStep*_PARAM(_timestepDiffusion));
 
     if (isDoseStartTime(realTime))
     {
